requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
docx-mailmerge==0.5.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
numpy==2.2.3
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-redmine==2.5.0
pytz==2025.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.4.0
webencodings==0.5.1
Werkzeug==3.1.3
python-docx
